Split off the drawer into own folder
